requirements.txt
python-telegram-bot==20.7
deep-translator==1.11.4
python-dotenv==1.0.0
langdetect==1.0.9
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0